refactor(auth): remove unnecessary getCsrfToken re-export from AuthContext
